vectorizer/pyproject.toml
[project]
name = "vectorizer-worker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "certifi>=2025.0.0",
]

[dependency-groups]
dev = [
  "pywrangler",
  "workers-py",
  "workers-runtime-sdk",
]

[tool.pywrangler]
